pyproject.toml
[project]
name = "news-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.1",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.3.6",
    "fastapi>=0.124.2",
    "langchain-chroma>=1.0.0",
    "langchain-google-genai>=4.0.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.38.0",
]
